Install react-select. Replace old tagselector with new component
