+++ packages/designmanual/stories/collated-components.jsx
@@ -21,6 +21,7 @@ import {
   OneColumn,
   Content,
   i18nInstance,
+  TagSelectorV2,
 } from '@ndla/ui';
 import Pager from '@ndla/pager';
 import { Translation as I18nTranslate } from 'react-i18next';
@@ -63,6 +64,7 @@ import MessageBox from './molecules/MessageBoxExample';
 import ResourceBoxExample from './pages/ResourceBoxExample';
 
 import TagSelectorExample from './molecules/TagSelectorExample';
+import TagSelectorV2Example from './molecules/TagSelectorExample';
 import SnackbarExample from './molecules/SnackbarExample';
 
 storiesOf('Sammensatte moduler', module)
