+++ packages/ndla-ui/package.json
@@ -58,6 +58,7 @@
     "lodash": "^4.17.20",
     "react-bem-helper": "1.4.1",
     "react-device-detect": "^2.1.2",
+    "react-select": "^5.4.0",
     "react-swipeable": "^6.2.0",
     "react-transition-group": "^2.5.3",
     "remarkable": "^2.0.1",
